requirements.txt
pandas
numpy
sentence-transformers
faiss-cpu
openai
streamlit
transformers  # for HuggingFace fallback
